PythonPlayGround/mcp_ppt/requirements.txt
fastmcp>=2.0.0
python-pptx>=0.6.21
pydantic>=2.0.0

